Version 3.30.7 (based on bleeding_edge revision r24503)
Fix computation of UTC time from local time at DST change points (issue 3116, Chromium issues 415424, 417640).

Convert `obj` ToObject in Object.keys() (issue 3587).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@24504 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mips64/macro-assembler-mips64.cc
@@ -6069,7 +6069,7 @@ void MacroAssembler::TruncatingDiv(Register result,
   DCHECK(!result.is(at));
   base::MagicNumbersForDivision<uint32_t> mag =
   base::SignedDivisionByConstant(static_cast<uint32_t>(divisor));
-  li(at, Operand(mag.multiplier));
+  li(at, Operand(static_cast<int32_t>(mag.multiplier)));
   Mulh(result, dividend, Operand(at));
   bool neg = (mag.multiplier & (static_cast<uint32_t>(1) << 31)) != 0;
   if (divisor > 0 && neg) {
